requirements.txt
pystray
Pillow
pydexcom
keyring
win10toast
tk
